requirements.txt
anyio==4.0.0
appdirs==1.4.4
APScheduler==3.10.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
configparser==6.0.0
crayons==0.4.0
cryptography==41.0.4
cssselect==1.2.0
dateparser==1.1.8
demjson3==3.0.6
dnspython==2.4.2
exceptiongroup==1.1.3
facebook-scraper==0.2.59
fake-useragent==1.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
kaitaistruct==0.10
lxml==4.9.3
outcome==1.2.0
parse==1.19.1
ptbcontrib @ git+https://github.com/python-telegram-bot/ptbcontrib.git@b3a6e3e2e5e0ed592944b3efc64b6162d0c90e08
pyasn1==0.5.0
pycparser==2.21
pyee==8.2.2
pymongo==3.12.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.6
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-html==0.10.0
selenium==4.1.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzlocal==5.1
urllib3==1.26.17
urllib3-secure-extra==0.1.0
w3lib==2.1.2
webdriver-manager==3.2.2
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
zstandard==0.21.0
mariadb==1.1.8